bot/requirements.txt
aiogram==3.0.0b7
ldap3==2.9.1
python-dotenv==1.0.0
async-timeout==4.0.3